The Great Code Massacre (delete lots of dead code)
Removed the following stubs from cl_init.lua:
- GM:Initialize (Empty hook)
- GM:Thing (Empty hook)
- GM:EntityRemoved (Empty hook)
- GM:PostDrawTranslucentRenderables (Empty hook)
- GM:DrawMonitors (Empty hook)
- GM:PreDrawTranslucentRenderables (Empty hook)
- GM:OnReloaded (Empty hook)
- GM:CreateMove (Empty hook)
- GM:PostDrawEffects (Empty hook)
- GM:CreateMove (Empty hook)
- GM:PostDrawEffects (Empty hook)
Removed the following stubs from init.lua:
- GM:ShutDown (Empty hook)
- GM:OnEndRound (Empty hook)
- GM:OnStartRound (Empty hook)
- GM:OnReloaded (Empty hook)
Removed cl_player.lua because the only code inside of it was the hook
GM:PlayerFootstep which had only commented code inside of it.
Removed some commented code in cl_rounds.lua because it was code from
mechanicalmind's melonbomber gamemode. Also simplified the if/else logic
in that section because, of the three branching cases, only one actually
did something.
Removed some code in cl_scoreboard.lua in the GM:DoScoreboardActionPopup
method because the code was commented out (it was code from
mechanicalmind's murder gamemode) and was not doing anything.
Removed a couple lines of code from sv_rounds.lua that were manually
calling hooks for OnEndRound and OnStartRound.