Reworked the prop banning system and menu
Props that are banned are no longer completely removed from the map.
Instead they are left in the map but are not highlighted as valid
objects to change into if the player mouses over the prop.
Removed the console commands for ph_bannedmodels_add and
ph_bannedmodels_remove. Their functionality has been replicated as net
messages.
The ph_bannedmodels_menu has been completely reworked.
- Banned props are now shown as an icon of their models instead of a
  line of text of their model string.
- If you are an admin you can hover over a prop icon and get a button to
  unban that particular prop. There is also a text field to ban new
  props at the top of the menu.
- Unbanned props become available to change into as soon as the same
  round in which they are unbanned. The same logic applies to props that
  are banned.
- The banned models menu now updates live as new net messages come in that
  modify the BannedModels list. This allows multiple players to have the
  menu open at the same time and see changes made to the list immediately.
Checks for whether or not a user is allowed to ban/unban a prop have
been moved serverside.
Banned props are now synced when the player connects to the server.