Reworked the chattext system
The chattext system is what allows the server to send colored messages
to all/individual players. It is needed because there is no existing
functionality provided by Garry's Mod for doing that.

The new system works similarly to the old system except that it has a
lot less bloat and is much easier to use/understand.

The chattext system was renamed to the chatmsg system
since that more accurately relfects what it does.