Added ui element showing map start wait time
Added a new ui element that shows how long until the game starts. This
is in reference to the initial period of time after a map first loads
where the game waits for players to join. The amount of time that the
game waits is controlled by the ConVar ph_mapstartwait. In order for
clients to have access to this ConVar its creation has been moved to
shared.lua and given the FCVAR_REPLICATED flag.