Added system for overriding default mapvote
Added in a system for overriding Prophunter's default mapvoting system.
The way it works is that in sv_mapvote.lua there are now two functions
that respectively identify and execute functionality for other mapvote
addons. By default the MapVote addon (Workshop ID 151583504) will be
compatible with Prophunters.
If the default mapvote is overridden then Prophunters will not do
anything in regards to mapvoting. This is so that it does not interfere
with the mapvote addon being used. Consequently, this also means that if
the mapvote addon fails to change the map then the gamemode as a whole
will break because Prophunters will wait an infinite amount of time for
the map to be changed. At this point the map must be manually changed or
server restarted.