Removed net sync and mapvote override helper
Removed the net sync of mapvote overrides because it actually wasn't
being used anywhere on the client.
Now checks for a mapvote override when the mapvoting phase of
Prophunters is triggered instead of at gamemode initialization.
Removed the override helper function because we're really only checking
for a single mapvote addon. If there ever ends up being more than one
then the helper can be added back.