Added map time limit config option (#10)
Added new convar ph_map_time_limit. This convar should be set to either
a positive integer or -1. A positive integer will enable the timer while
a value of -1 will disable the timer. The timer is disabled by default.
If a positive integer is given it will be interpreted as the number of
minutes a map should last. When the timer expires it will declare the
following round to be the last round.

Added a chat message that shows when the current round is the last
round.