Added sv_awards file for managing endround awards
The sv_awards.lua file is new system for managing the end of round
player awards. All of the award types in prophunters have been moved to
this file. This file easily supports adding new rewards by simply
following the format of the awards already in the file.
This commit simply checks in this file, it does not modify the existing
code to use this new system.